requirements.txt
annotated-types==0.5.0
anyio==3.7.1
black==23.7.0
cffi==1.15.1
click==8.1.5
colorama==0.4.6
cryptography==41.0.2
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pytest==7.4.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.22.0
